requirements.txt
numpy==1.24.3
torch==2.0.1
matplotlib==3.7.2
tqdm==4.65.0
langchain>=0.0.267
langchain-core>=0.0.10
langchain-openai>=0.0.2
langchain-anthropic>=0.0.2
langchain-community>=0.0.2
langchain-deepseek>=0.0.1
pydantic>=2.0.0
python-dotenv>=1.0.0 